+++ src/snippets/common.snippet.ts
@@ -4,7 +4,7 @@
  * Created Date: 2023/12/17 Sun 10:40:38
  * Author: Donghee Kim (terdong@gmail.com)
  * -----
- * Last Modified: 2023/12/17 Sun 11:45:16
+ * Last Modified: 2024/01/04 Thu 07:52:48
  * Modified By: Donghee Kim (terdong@gmail.com)
  * -----
  * Copyright (c) 2023 TeamGehem
@@ -14,8 +14,8 @@
  * ----------	---	---------------------------------------------------------
  */
 
-export const getCommonSnippet: (packagePath: string, name: string) => string = (packagePath, name) => `
-package ${packagePath}
+export const getCommonSnippet: (packagePath: string, name: string) => string = (packagePath, name) =>
+  `package ${packagePath}
 
 object ${name}: