+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "DDD File Generator" extension will be documented in this file.
 
+## [1.1.0] (2024-01-04)
+
+- Fix command name
+- Add another command for it without files
+- Add configuration. Now It can be specified the default root path and the directory name of each layer in setting
+
 ## [1.0.1] (2023-12-18)
 
 - Hotfix